# smarthire-backend/specs/001-smarthire-backend-platform/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: SmartHire Backend API
  description: |
    Internal recruitment management platform API.
    All endpoints require a valid JWT bearer token except `/prescreen/**`, `/health`, `/docs`, and `/openapi.json`.
  version: "1.0.0"
  contact:
    name: SmartHire Platform Team
servers:
  - url: "http://localhost:8083"
    description: Local development
  - url: "https://smarthire.internal"
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ResponseDto:
      type: object
      properties:
        response:
          type: array
          items: {}
        message:
          type: string
        exception:
          type: string
          nullable: true
        result:
          type: string
          nullable: true
      required:
        - response
        - message
...
    InterviewSlotRequest:
      type: object
      required: [interviewerCalendarId, candidateId, recruiterId]
      properties:
        interviewerCalendarId:
          type: integer
          format: int64
        candidateId:
          type: integer
          format: int64
        recruiterId:
          type: integer
          format: int64
        interviewTypeId:
          type: integer
          format: int64
...
    RescheduleRequest:
      type: object
      required: [interviewerCalendarId, newDate, newStartTime]
      properties:
        interviewerCalendarId:
          type: integer
          format: int64
        newDate:
          type: string
          format: date
        newStartTime:
          type: string
          example: "14:00"

    # -------------------------------------------------------------------------
    # Feedback Form
    # -------------------------------------------------------------------------
    FeedbackFormCreateRequest:
      type: object
      required: [templateName, technologyId, nodes]
      properties:
        templateName:
          type: string
        technologyId:
          type: integer
          format: int64
        practiceId:
          type: integer
          format: int64
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackFormNode"
    FeedbackFormNode:
      type: object
      required: [heading]
      properties:
        heading:
          type: string
        parentId:
          type: integer
          format: int64
          nullable: true
        dataTypeId:
          type: integer
          format: int64
        requiredFlag:
          type: boolean
          default: false
        orderIndex:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/FeedbackFormNode"
    FeedbackFormHeadingsRequest:
      type: object
      required: [technologyId]
      properties:
        technologyId:
          type: integer
          format: int64
        practiceId:
          type: integer
          format: int64

    # -------------------------------------------------------------------------
    # User / Employee
    # -------------------------------------------------------------------------
...
    MeetingLinkRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: OAuth2 authorization code from /getCode redirect
    MeetingInviteRequest:
      type: object
      required: [candidateId, interviewerCalendarId]
      properties:
        candidateId:
          type: integer
          format: int64
        interviewerCalendarId:
          type: integer
          format: int64
        subject:
          type: string
        attendeeEmails:
          type: array
          items:
            type: string
            format: email

    # -------------------------------------------------------------------------
    # Keycloak
    # -------------------------------------------------------------------------
...
paths:
...
  /recruiter/getDirectRecruiterSlots:
    post:
      tags: [Recruiter]
      summary: Retrieve interview slots for a recruiter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recruiterEmail:
                  type: string
                  format: email
                month:
                  type: integer
                year:
                  type: integer
      responses:
        "200":
          description: Recruiter slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /recruiter/saveInterviewSlot:
    post:
      tags: [Recruiter]
      summary: Book a direct interview slot (without prior availability slot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewSlotRequest"
      responses:
        "200":
          description: Slot booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /recruiter/rescheduleSlot:
    post:
      tags: [Recruiter]
      summary: Reschedule a recruiter-created interview slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleRequest"
      responses:
        "200":
          description: Slot rescheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /recruiter/deleteSlot:
    post:
      tags: [Recruiter]
      summary: Delete a recruiter-created interview slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recruiterCalendarId]
              properties:
                recruiterCalendarId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Slot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /recruiter/feedbackForm:
    post:
      tags: [Recruiter]
      summary: Retrieve feedback form for a specific interview slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interviewerCalendarId]
              properties:
                interviewerCalendarId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: Feedback form data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /recruiter/feedbackFormPDF:
    post:
      tags: [Recruiter]
      summary: Generate or retrieve the PDF feedback report for an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [interviewerCalendarId]
              properties:
                interviewerCalendarId:
                  type: integer
                  format: int64
      responses:
        "200":
          description: PDF URL or file stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Slot Bulk Upload
  # ===========================================================================
...
  /feedbackForm/addFeedbackForm:
    post:
      tags: [FeedbackForm]
      summary: Create or update a feedback form template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackFormCreateRequest"
      responses:
        "200":
          description: Template saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /feedbackForm/getFeedbackFormHeadings:
    post:
      tags: [FeedbackForm]
      summary: Retrieve active hierarchical feedback form for a technology and practice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackFormHeadingsRequest"
      responses:
        "200":
          description: Feedback form hierarchy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Reports
  # ===========================================================================
...
  /getCode:
    get:
      tags: [Teams]
      summary: Initiate Microsoft OAuth2 authorization code flow
      responses:
        "302":
          description: Redirect to Microsoft login
        "200":
          description: Authorization code returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /getMeetingLink:
    post:
      tags: [Teams]
      summary: Exchange auth code for access token and create Teams meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingLinkRequest"
      responses:
        "200":
          description: Teams meeting join URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /sendMeetingInvite:
    post:
      tags: [Teams]
      summary: Create Teams meeting and send invite to attendees
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInviteRequest"
      responses:
        "200":
          description: Meeting invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Alerts
  # ===========================================================================
...
  /referralCandidate/submitReferral:
    post:
      tags: [Referral]
      summary: Submit an employee referral
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [referrerEmployeeId, candidateName, emailId]
              properties:
                referrerEmployeeId:
                  type: string
                candidateName:
                  type: string
                emailId:
                  type: string
                  format: email
                mobileNumber:
                  type: string
                totalExp:
                  type: number
                  format: float
                currentCtc:
                  type: number
                  format: double
                noticePeriod:
                  type: string
                buId:
                  type: integer
                  format: int64
                technologyIds:
                  type: array
                  items:
                    type: integer
                    format: int64
                certificationIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        "200":
          description: Referral submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"
  /referralCandidate/bulkUpload:
    post:
      tags: [Referral]
      summary: Bulk upload referral candidates via Excel
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Upload result with row-level errors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseDto"

  # ===========================================================================
  # Supply Screen
  # ===========================================================================
...